SM-815: include patches into CyclonDx SBOM file generated by yocto (#6)
The following was changed:
- Do not upload empty SBOM every time build starts
- Created multiple python files to store more logic instead of storing
everything in 1 bbclass file
- After the SBOM is finished in the upload step the build now checks if
the project exists and creates a clone on the latest version (which
stores all vulnerabilities + analysis). The clone operation takes some
time, but is relatively fast (experiments show up to a minute) so the
build would wait until the operation is done checking the token every 10
seconds

Known issues:
- the init and upload step is run twice. Not causing a lot of issues as
of now, might even help with dependency track not finding some
vulnerabilities fast enough. Since clone checks if the project exists
the second time it is not triggered
- Based on dependency track behavior in case a new component is added
into a new version it is still possible that it will miss a component
due to sync issue. InternalAnalysisTask's token is not available to see
if BOM analysis is finished, but cloning should help a lot with the
vulnerabilities that were missed